Rewrite code generation to guard against undefined
Due to strange circumstances, code generation didn't guard against
undefined objects.

That is, we depended on the value of the variable
_atLeastOneSpecialIterator for guarding: when it was false, we didn't
guard; when it was true, we guarded. However, the variable was
initialized at false and remained unchanged.

In this new version, we ignore this variable completely and guard
against undefined objects no matter what.