Add support for dynamic keys in objects
This patch adds support for referring to arbitrary keys in objects using
variables, such as "item.names.$$language". It provides an alternative to
passing arrays of objects for cases when the input data is a flat
key-value map, or when the object is complex and the key to get
data from depends on some other data (currencies, environment, etc.).