Store template image width/height in a immutable variable to avoid weird bugs
Before, imageHeight/imageWidth contained the image size of the template and  also the final image size at the end of the process. The same attributes was used for the two different concepts
If two XML files are parsed one after another with the same image, the second XML file overwrite the previous result. It creates a bug.