fix: `:html` formatter did not remove forbidden characters such as `\u0010` for DOCX/ODT...
Html strings are injected after the builder  in `assembleXmlPart`. So the builder does not remove forbidden characters in HTML
I could apply this correction in `assembleXmlPart` but at the moment, only html injection is concerned. So I prefer to remove these characters only in the html system for performance reasons.